refactor: extract shared scanning utilities from extract.ts and redact.ts
findPhoneCandidates and findCreditCardCandidates in redact.ts were
byte-for-byte identical scan loops, differing only in which
character Set they checked. stripTrailingPunctuation (extract.ts)
and stripTrailingDots (redact.ts) were the same trim-loop pattern
too. Both now share a new internal module, src/text/internal/scan.ts
(scanMaximalRuns, stripTrailing).

findEmailCandidates and findUrlCandidates are deliberately left
untouched -- they're structurally different (anchor-based @-split
and prefix-indexOf scanning, not homogeneous character-class runs),
so forcing them into the same abstraction would trade real
duplication for a worse, over-general one.

No behavior change: all 29 existing extract/redact tests pass
unchanged, and the ReDoS benchmarks confirm linear-time performance
is preserved.

Closes #342
